Add parameterize back date in app.config

+++ AttendanceAutoRun/Program.cs
@@ -15,29 +15,40 @@ namespace AttendanceAutoRun
 
         static void Main(string[] args)
         {
+            var ErrorStts = false;
+            if (ConfigurationManager.AppSettings["DebugConsole"].ToLower() == "true")
+                ErrorStts = true;
             try
             {
+
                 con = new OleDbConnection();
                 //con.ConnectionString = "Provider=Microsoft.ACE.OLEDB.12.0;Data Source=D:/Private/Coba-coba/attBackup.mdb";
                 con.ConnectionString = ConfigurationManager.AppSettings.Get("AccessConnectionString");
                 cmd = new OleDbCommand();
                 cmd.Connection = con;
-                var whereDate = DateTime.Today.Date;
+
+                int backDay = 1;
+                var cek  = Int32.TryParse(ConfigurationManager.AppSettings["BackDay"], out backDay);
+                backDay = !cek ? 1 : backDay;
+
+                var whereDate = DateTime.Today.AddDays(-backDay).Date;
                 var conn = ConfigurationManager.AppSettings.Get("SQLServerConnectionString");
-                var item = 0;
+                //var item = 0;
+                var item = false;
 
                 using (SqlConnection connection = new SqlConnection(conn))
                 {
                     connection.Open();
                     using (var cmd = connection.CreateCommand())
                     {
-                        cmd.CommandText = "Select count(*) from CHECKINOUT";
-                        item = (Int32)cmd.ExecuteScalar();
+                        //cmd.CommandText = "Select count(*) from CHECKINOUT";
+                        cmd.CommandText = "select case when count(*) > 0 then 1 else 0 end from CHECKINOUT";
+                        item = (Int32)cmd.ExecuteScalar() > 0;
                     }
                     connection.Close();
                 }
                 var query = "SELECT * FROM CHECKINOUT";
-                query += item < 1 ? "" : " where CHECKTIME >= #" + whereDate.ToString("dd/MM/yyyy HH:mm:ss") + "#";
+                query += !item ? "" : " where CHECKTIME >= #" + whereDate.ToString("dd/MM/yyyy HH:mm:ss") + "#";
                 
                 cmd.CommandText = query;
                 con.Open();
@@ -50,11 +61,14 @@ namespace AttendanceAutoRun
                 {
                     connection.Open();
 
-                    using (var cmd = connection.CreateCommand())
+                    if(item)
                     {
-                        cmd.CommandText = "DELETE FROM dbo.CHECKINOUT WHERE CHECKTIME >= @word";
-                        cmd.Parameters.AddWithValue("@word", whereDate);
-                        cmd.ExecuteNonQuery();
+                        using (var cmd = connection.CreateCommand())
+                        {
+                            cmd.CommandText = "DELETE FROM dbo.CHECKINOUT WHERE CHECKTIME >= @word";
+                            cmd.Parameters.AddWithValue("@word", whereDate);
+                            cmd.ExecuteNonQuery();
+                        }
                     }
 
                     using (SqlBulkCopy bulkCopy = new SqlBulkCopy(connection))
@@ -68,6 +82,7 @@ namespace AttendanceAutoRun
                         }
                         catch (Exception ex)
                         {
+                            ErrorStts = true;
                             Console.WriteLine("==================================== Filed Execute ====================================");
                             Console.WriteLine("Exception : " + ex.Message);
                         }
@@ -80,11 +95,11 @@ namespace AttendanceAutoRun
                         cmd.ExecuteNonQuery();
                     }
 
-                    //var queryAttd = ConfigurationManager.ConnectionStrings["QueryToAttendanceMachinePolling"].ConnectionString;
+                    var queryAttd = ConfigurationManager.AppSettings["MigrationQuery"].ToLower();
                     using (var cmd = connection.CreateCommand())
                     {
-                        cmd.CommandText = "insert into AttendanceMachinePolling select null as Id, SENSORID as Barcode, concat(cast(CHECKTIME as date), ' 00:00:00.000') as AttendanceDate, Cast(CHECKTIME as time(7)) as AttendanceTime, case when CHECKTYPE = 'o' then 0 else 1 end as AttendanceType, GETDATE() as CreatedDate from CHECKINOUT";
-                        //cmd.Parameters.AddWithValue("@word", whereDate);
+                        cmd.CommandText = queryAttd;
+                        //cmd.CommandText = "insert into AttendanceMachinePolling select null as Id, cast([USERID] as nvarchar()) as Barcode, concat(cast(CHECKTIME as date), ' 00:00:00.000') as AttendanceDate, Cast(CHECKTIME as time(7)) as AttendanceTime, case when CHECKTYPE = 'o' then 0 else 1 end as AttendanceType, GETDATE() as CreatedDate from CHECKINOUT";
                         cmd.ExecuteNonQuery();
                     }
                 }
@@ -94,11 +109,12 @@ namespace AttendanceAutoRun
             }
             catch (Exception e)
             {
+                ErrorStts = true;
                 Console.WriteLine("==================================== Filed Execute ====================================");
                 Console.WriteLine("Exception : " + e.Message);
             }
-
-            Console.ReadKey();
+            if(ErrorStts)
+                Console.ReadKey();
         }
     }
 }
